historical-search-api/requirements.txt
Flask-Caching==1.10.1
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-Moment==0.11.0
Flask-SQLAlchemy==2.5.1
Flask-Script==2.0.5
Flask==2.2.5
Flask-OpenTracing==1.1.0
Jinja2==3.0.3
Mako==1.2.2
MarkupSafe==2.1.4
SQLAlchemy-Continuum==1.3.11
SQLAlchemy-Utils==0.36.8
SQLAlchemy==1.3.24
Werkzeug==2.3.8
alembic==1.5.8
aniso8601==9.0.1
asyncio-nats-client==0.11.4
asyncio-nats-streaming==0.4.0
attrs==20.3.0
bcrypt==3.2.0
blinker==1.4
boto3==1.24.35
cachelib==0.1.1
certifi==2023.7.22
cffi==1.14.5
chardet==4.0.0
click==8.1.3
flask-jwt-oidc==0.3.0
flask-marshmallow==0.11.0
flask-restplus
flask_restx
flask-cors==3.0.10
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
jaeger-client==4.4.0
jaro-winkler==2.0.3
jsonschema==3.2.0
marshmallow-sqlalchemy==0.23.1
marshmallow==3.0.0rc7
minio==7.0.2
opentracing==2.4.0
protobuf==3.18.3
psycopg2-binary==2.8.6
pycparser==2.20
pyhumps==1.6.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.16.0
python-editor==1.0.4
python-jose==3.2.0
pytz==2021.1
requests==2.31.0
rsa==4.7.2
sentry-sdk==1.14.0
six==1.15.0
threadloop==1.0.2
thrift==0.13.0
tornado==6.1
flask-expects-json==1.5.0
redis==4.1.4
flask_jwt_oidc==0.3.0
maya==0.6.1
pyjwt==2.4.0
aws-requests-auth==0.4.3
holidays==0.12
Flask-SocketIO==5.3.6
Flask-Login==0.5.0
eventlet==0.35.2
uritemplate.py==3.0.2
urllib3==1.26.15
ndg-httpsclient
pyopenssl
pyasn1
secure==0.3.0
boto3==1.24.35
walrus==0.8.2
imap-tools==0.56.0
more-itertools==10.2.0
-e git+https://github.com/bcgov/sbc-common-components.git#egg=sbc-common-components&subdirectory=python